+++ src/profile.ts
@@ -1,21 +1,6 @@
 /// Profile
 
-(() => {
-	window.addEventListener("message", async (event) => {
-		const { source, endpoint, data } = event.data;
-		if (source !== "lavaplace" || !endpoint) return;
-		
-		const endpointText = endpoint.split('?')[0].split('/').filter((s: string) => s && isNaN(Number(s))).filter((s: string) => s && !s.includes('.')).pop();
-
-		switch (endpointText) {
-			case "me":
-				await updateUserMenu();
-				break;
-		}
-	});
-})();
-
-async function updateUserMenu() {
+async function updateUserMenu(data?: meData) {
 	let userButton = document.querySelector(".dropdown .btn[title=\"Show profile\"]"); // English
 	if (!userButton) userButton = document.querySelector(".dropdown .btn[title=\"Exibir perfil\"]"); // Português
 
@@ -25,8 +10,23 @@ async function updateUserMenu() {
 	if (!menu) return false;
 	if (!menu.id) menu.id = "user-menu";
 
+	fixUserMenu(menu);
 	generateLevelInfo(menu);
+	generateSelfUserCopy(menu, data);
+	await generateAllianceInfo(menu, data);
+
 	return true;
+}
+
+function fixUserMenu(menu: Element) {
+	const avatar = menu.querySelector("section .avatar");
+	if (avatar) {
+		const editProfileButton = avatar.parentElement?.querySelector("button");
+		if (editProfileButton) {
+			(editProfileButton as HTMLButtonElement).classList.remove("-bottom-1");
+			(editProfileButton as HTMLButtonElement).style.top = `${avatar.getBoundingClientRect().height - editProfileButton.getBoundingClientRect().height}px`;
+		}
+	}
 }
 
 function getNextLevelPixels(menu: Element) {
@@ -54,7 +54,7 @@ function generateLevelInfo(menu: Element) {
 
 	const div = document.createElement("div");
 	div.className = "flex items-center gap-1";
-	div.id = "lava-place-level-info";
+	div.id = "lavaplace-level-info";
 
 	const iconPath = document.createElementNS("http://www.w3.org/2000/svg", "path");
 	iconPath.setAttribute("d", "m 160,-520 h 487 l -224,-224 57,-56 320,320 -320,320 -57,-56 224,-224 H 160 Z");
@@ -76,6 +76,87 @@ function generateLevelInfo(menu: Element) {
 	div.appendChild(icon);
 	div.appendChild(text);
 
-	if (topSection[1].querySelector("#lava-place-level-info")) topSection[1].querySelector("#lava-place-level-info")?.replaceWith(div);
+	if (topSection[1].querySelector("#lavaplace-level-info")) topSection[1].querySelector("#lavaplace-level-info")?.replaceWith(div);
 	else topSection[1].appendChild(div);
+}
+
+function removeAllianceInfo(menu: Element) {
+	const allianceInfo = menu.querySelector("#lavaplace-alliance-info");
+	if (allianceInfo) allianceInfo.remove();
+}
+
+function generateSelfUserCopy(menu: Element, data?: meData) {
+	const topSection = menu.querySelectorAll("section > div");
+	if (!topSection || topSection.length < 2) return;
+
+	const userIdSpan = topSection[1].querySelector("div.items-center.font-medium > span");
+	if (!userIdSpan) return;
+
+	let user = {
+		name: data?.name,
+		id: data?.id,
+	}
+
+	if (!user.name || !user.id) {
+		user.id = Number(userIdSpan.textContent.match(/#(\d+)/)?.[1]);
+		user.name = topSection[1].querySelector("h3")?.textContent;
+	}
+
+	console.log('user:', user);
+
+	if (!user.name || !user.id) return;
+
+	const copy = `My user name: ${user.name} (ID: #${user.id})`;
+
+	(userIdSpan as HTMLSpanElement).style.cursor = "pointer";
+	showTooltip(userIdSpan as HTMLSpanElement, "Copy User");
+
+	(userIdSpan as HTMLSpanElement).onclick = () => {
+		GM_setClipboard(copy, "text");
+		showTooltip(userIdSpan as HTMLSpanElement, "Copied User!", 1500);
+	};
+}
+
+function generateSelfAllianceCopy(element: Element, alliance: allianceData) {
+	const copy = `My alliance name: ${alliance.name} (ID: &${alliance.id})`;
+	
+	(element as HTMLSpanElement).style.cursor = "pointer";
+	showTooltip(element as HTMLSpanElement, "Copy Alliance");
+
+	(element as HTMLSpanElement).onclick = () => {
+		GM_setClipboard(copy, "text");
+		showTooltip(element as HTMLSpanElement, "Copied Alliance!", 1500);
+	};
+}
+
+async function generateAllianceInfo(menu: Element, data?: meData) {
+	if (data && data.allianceId <= 0) return removeAllianceInfo(menu);
+
+	const topSection = menu.querySelectorAll("section > div");
+	if (!topSection || topSection.length < 2) return removeAllianceInfo(menu);
+
+	const userElement = topSection[1].querySelector("div.items-center");
+	if (!userElement) return removeAllianceInfo(menu);
+
+	let alliance: allianceData | undefined;
+
+	try {
+		const res = await fetch("https://backend.wplace.live/alliance", { credentials: "include" });
+		if (res.ok) alliance = await res.json();
+	} catch (error) {
+		console.debug(error);
+	}
+
+	if (!alliance || !alliance.id || alliance.id <= 0) return removeAllianceInfo(menu);
+
+	const span = document.createElement("span");
+	span.id = "lavaplace-alliance-info";
+	span.innerText = alliance.name;
+	span.className = "badge badge-sm ml-0.5 border-0";
+	span.style.backgroundColor = "color-mix(in oklab, var(--color-primary) 10%, transparent)";
+	span.style.color = "var(--color-primary)";
+	span.style.marginBottom = "5px";
+
+	topSection[1].insertBefore(span, topSection[1].children[1]);
+	generateSelfAllianceCopy(span, alliance);
 }