+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v1.1
+
+- Added the user's alliance in profile
+- Added a way to copy user & alliance IDs (pixel infos & user's profile)
+
 ### v1.0
 
 - Initialized the repository